Updates to match public API in review
