+++ Example/CryptoOnramp Example/CryptoOnramp Example/RegistrationView.swift
@@ -95,11 +95,7 @@ struct RegistrationView: View {
 
                 if let customerId = registrationCustomerId {
                     HiddenNavigationLink(
-                        destination: AuthenticatedView(
-                            coordinator: coordinator,
-                            customerId: customerId,
-                            email: email
-                        ),
+                        destination: AuthenticatedView(coordinator: coordinator, customerId: customerId),
                         isActive: $showAuthenticatedView
                     )
                 }
