+++ StripeCryptoOnramp/StripeCryptoOnramp/Source/Components/CryptoOnrampCoordinator.swift
@@ -66,12 +66,10 @@ public protocol CryptoOnrampCoordinatorProtocol {
     /// - Parameter network: The crypto network for the wallet address.
     func collectWalletAddress(walletAddress: String, network: CryptoNetwork) async throws
 
-    /// Presents the user with the ability to select a payment method.
-    /// - Parameters:
-    ///   - viewController: The view controller from which to present the payment method selector.
-    ///   - email: The email address to pre-fill in the presented sheet. If `nil`, the email field will be empty.
-    /// - Returns: The result of the selection. Returns `completed` if the user selected a payment method with an associated `PaymentMethodPreview` describing the selected method, or `canceled` if they didn’t complete selection.
-    func presentPaymentMethodSelector(from viewController: UIViewController, email: String) async -> PaymentMethodSelectionResult
+    /// Presents the Link sheet to collect a customer's payment method.
+    /// - Parameter viewController: The view controller from which to present the Link sheet.
+    /// - Returns: A `PaymentMethodPreview` if the user selected a payment method, or `nil` otherwise.
+    func collectPaymentMethod(from viewController: UIViewController) async -> PaymentMethodPreview?
 }
 
 /// Coordinates headless Link user authentication and identity verification, leaving most of the UI to the client.
@@ -203,11 +201,12 @@ public final class CryptoOnrampCoordinator: CryptoOnrampCoordinatorProtocol {
         )
     }
 
-    public func presentPaymentMethodSelector(from viewController: UIViewController, email: String) async -> PaymentMethodSelectionResult {
+    public func collectPaymentMethod(from viewController: UIViewController) async -> PaymentMethodPreview? {
+        let email = try? await linkAccountInfo.email
         if let result = await linkController.collectPaymentMethod(from: viewController, with: email) {
-            return .completed(.init(icon: result.icon, label: result.label, sublabel: result.sublabel))
+            return PaymentMethodPreview(icon: result.icon, label: result.label, sublabel: result.sublabel)
         } else {
-            return .canceled
+            return nil
         }
     }
 }
